\name{blimatesting}
\alias{blimatesting}
\docType{data}
\title{blimatesting object for testing purposes of blima package}
\description{This object is derived from the dataset of experiment on
human mesenchymal cells. Experiment was performed using two
Illumina arrays HumanHT-12 v4 Expression BeadChip Kit with
total 24 array spots. For the testing purposes only 9 conditions were selected to be included to this object due to space reasons. We name the arrays used for downstream analysis A1,A2, A3, A4 for condition A (Group A) (cells grown in alfa-MEM medium with 10\% fetal bovine serum) and E1, E2, E3, E4 for condition E (Group E)
(cells grown in CellGro medium, with human serum and suplements
FGF-2, EGF, M-CSF and insulin). In the set there is also included array labeled D4 for condition D(cells grown in CellGro medium, with human serum and suplements PDGF-BB, EGF, M-CSF and insulin).

These data has been deposited to NCBI Gene Expression Omnibus site as \href{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE56129}{GSE56129}.
}
\usage{data(blimatesting)}
\keyword{datasets}
\author{Vojtech Kulvait}
